requirements.txt
bresenham==0.2.1
cachetools==5.5.0
colorama==0.4.6
einops==0.8.1
flatten_dict==0.4.2
gdown==4.6.0
h5py==3.11.0
lpips==0.1.4
matplotlib==3.7.4
networkx==3.1
omegaconf==2.3.0
opencv_python==4.9.0.80
pandas==2.0.3
pillow==10.2.0
plotly==6.0.0
positional_encodings==6.0.3
pycocotools==2.0.7
pylcs==0.1.1
pytesseract==0.3.13
python_doctr==0.7.0
python_Levenshtein==0.12.2
pytorch_lightning==2.1.3
pytorch_msssim==0.2.1
PyYAML==6.0.2
rasterio==1.3.9
Requests==2.32.3
scikit-image==0.21.0
scikit_learn==1.3.2
scipy==1.9.3
Shapely==2.0.7
singleton_decorator==1.0.0
timm==0.5.4
tqdm==4.66.1
tensorboard==2.14.0
unflatten==0.1.1
